Support parallel read/parse for JSON parsers.
This involves refactoring the parse functions as methods of a functor-like
class, so that I don't need to add a parallel= argument to each overload. I
also renamed some of the overloads to use different function names for clarity.

Updated the dependencies to get byteme's parallel reader. Also added some
more tests to check some of the new methods.