Support a scalar representation for length-1 vectors. (#2)
Other languages no longer need to box/unbox their scalars when going to/from an
uzuki2-formatted file. Of course, R will continue to box them as it doesn't
make a distinction between scalars and length-1 vectors, but at least we get a
more faithful roundtrip in Javascript and Python.