Added versioning to accommodate factor, date(time) changes. (#5)
The specification now has a place to store the version inside the file, so
validators can do version-specific tasks. This is motivated by the deprecation
of ordered, date and date- types in favor of simplifying the type system.  Now,
ordered type is just a factor with ordered=true, while the date(time)s are just
strings with a special format. This should simplify interpretation a lot.