Removed the date, datetime vectors from the interface.
These are no longer necessary as they are just minor string variants.
We modify the string vector constructor interface to accept the format,
so applications can specialize the representation during initialization.

We also explicitly list out the interfaces rather than relying on a
TypedVector subclass, to get rid of an unnecessary inheritance layer.