More rigorous HDF5 support for missing values via placeholders. (#6)
This uses the same approach as the string vectors, where the data generator can
specify a placeholder for missing values (which doesn't overlap any of the
actual non-missing values, obviously). We no longer need to count on hard-coded
special values, in particular -2^31 for integers/booleans or R's special NA
double; this makes it a lot safer to transfer NAs between languages.