Read HDF5 datasets by block for memory efficiency. (#10)
This avoids loading in the entire dataset in a large temporary buffer;
rather, we load in blocks (i.e., multiples of chunks) until we process
the entire dataset. Currently the block size is hard-coded to process
up to 10000 elements at a time; this is probably fine for now.

Also needed to update the Action to get everything to pass.