Align with the new HDF5 policy for missing values, datatypes. (#13)
In version 1.3, number vectors are now allowed to be represented as HDF5
integer datatypes, provided the integer can be fully represented by a
double-precision float (i.e., 32-bit or lower).

If a missing value placeholder for a number is NaN, all NaNs are to be
considered missing. We no longer consider the NaN payloads.

Also updated the various ritsuko calls for better error messages.