Throw an error if the top-level object is not a list.
This makes sense as uzuki2 is, after all, designed to load lists, not assorted
vectors or factors. Nonetheless, it can be disabled with a new 'strict_list'
option if the caller doesn't mind dealing with non-list outputs.

This option requires a new class for the HDF5 parsers. While we're making
such a class, we took the opportunity to allow users to modify the HDF5 buffer
size, rather than hard-coding it into the various functions.