Programmatically generate version-specific specification documents.
This is arguably easier to read if we want to understand any given version of
the spec, as we don't have to mentally ignore the parts of the spec related to
other versions. We  use knitr to generate one document per version, only
keeping the clauses relevant to that version via conditional chunks.